packages/core/pyproject.toml
[tool.poetry]
name = "inbound-core"
version = "0.0.1"
description = "Inbound Core"
authors = [ "Virksomhetsdatalaget <virksomhetsdatalaget@nav.no>", ]
readme = "README.md"
packages = [{ include = "inbound" }]

[tool.poetry.dependencies]
python = "^3.10.0"
pandas = "2.0.0"
loguru = "^0.7.0"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"
google-cloud-secret-manager = "2.16.1"
google-auth = "2.17.3"
click = "^8.1.3"
pygit2 = "^1.12.0"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
bump-pydantic = "^0.6.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^4.0.1"
pyright = "^1.1.303"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
jupyter = "1.0.0"

[tool.poetry.scripts]
inbound = "inbound.core.cli:inbound"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


